Pet data fetch to parent, Like in Modal fixed
- elevated fetching of pet data to Matching (parent) component
- fixed 'Like' status updating in Modal & vice versa
