+++ frontend/src/components/Photocards.js
@@ -1,6 +1,6 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import { format } from "date-fns";
-import { lightBlue, pink, grey } from '@mui/material/colors';
+import { lightBlue } from '@mui/material/colors';
 import PetsRoundedIcon from '@mui/icons-material/PetsRounded';
 import WcRoundedIcon from '@mui/icons-material/WcRounded';
 import TodayRoundedIcon from '@mui/icons-material/TodayRounded';
@@ -8,39 +8,15 @@ import MonitorWeightRoundedIcon from '@mui/icons-material/MonitorWeightRounded';
 import EventAvailableRoundedIcon from '@mui/icons-material/EventAvailableRounded';
 import FavoriteBorderIcon from '@mui/icons-material/FavoriteBorder';
 import FavoriteIcon from '@mui/icons-material/Favorite';
-import CloseRoundedIcon from '@mui/icons-material/CloseRounded';
 import './photocards.css';
 import Modal from './Modal';
 
-// Function to return true if pet's Availability is "Available"
-const filterAvailable = (pets) => {
-    return pets.petAvailability === "Available";
-}
 
-export default function Photocard() {
-    // fetch and store full pet data
-    const [fullPetData, setFullPetData] = useState([]);
-    useEffect(() => {
-        fetch('http://localhost:5000/pets')
-        .then((res) => {
-            return res.json();
-        })
-        .then((data) => {
-            console.log(data);
-            setFullPetData(data);
-        })
-        .catch(function(error) {
-            console.log('Request failed', error);
-            window.location.reload(false);
-        });
-    }, []);
-
-    // array of pets (objects) that are "Available" for display on page
-    const availablePets = fullPetData.filter(filterAvailable);
-
+export default function Photocard({ availablePets }) {
+    // like pets functionality
     const [likedPets, setLikedPets] = useState([]);
 
-    const toggleLike = (petID) => {
+    function toggleLike(petID) {
         if (likedPets.includes(petID)) {
             setLikedPets(likedPets.filter(id => id !== petID));
         } else {
@@ -104,16 +80,11 @@ export default function Photocard() {
                 <div className="modal">
                     <div onClick={toggleModal} className="overlay"></div>
                     <div className="modal-content">
-                        <Modal pet={availablePets.find(p => p.petID === pet)} />
-                        <div className="modal-btns">
-                            <button className="close-modal" onClick={toggleModal}>
-                                <CloseRoundedIcon sx={{color: grey[900], fontSize: 36 }} />
-                            </button>
-
-                            <button className="like-modal" onClick={() => toggleLike(pet.petID)}>
-                                {likedPets.includes(pet.petID) ? <FavoriteIcon sx={{color: pink[500], fontSize: 36 }} /> : <FavoriteBorderIcon sx={{color: pink[500], fontSize: 36 }}/>} 
-                            </button>
-                        </div>
+                        <Modal 
+                            likedPets={likedPets}
+                            toggleLike={toggleLike}
+                            toggleModal={toggleModal}
+                            pet={availablePets.find(p => p.petID === pet)} />
                     </div>
                 </div>
             )}
