+++ frontend/src/pages/Matching.js
@@ -1,9 +1,33 @@
-import { useState } from "react";
+import { useState, useEffect } from "react";
 import Matchfilter from '../components/Matchfilter';
 import Photocards from '../components/Photocards';
 import './matching.css';
 
+// Function to return true if pet's Availability is "Available"
+const filterAvailable = (pets) => {
+    return pets.petAvailability === "Available";
+}
+
 export default function Matching() {
+    // fetch and store full pet data
+    const [fullPetData, setFullPetData] = useState([]);
+    useEffect(() => {
+        fetch('http://localhost:5000/pets')
+        .then((res) => {
+            return res.json();
+        })
+        .then((data) => {
+            console.log(data);
+            setFullPetData(data);
+        })
+        .catch(function(error) {
+            console.log('Request failed', error);
+            window.location.reload(false);
+        });
+    }, []);
+
+    // array of pets (objects) that are "Available" for display on page
+    const availablePets = fullPetData.filter(filterAvailable);
     
     const [isFilter, setIsFilter] = useState({
         0: [],    // type
@@ -54,7 +78,8 @@ export default function Matching() {
             </div>
             
             <div className="matching-main">
-                <Photocards />
+                <Photocards 
+                    availablePets={availablePets} />
             </div>
         </>
     )
